refactor(ai): rename chat.tsx to ai-search.tsx, remove re-exports
- Rename ai/chat.tsx → ai/ai-search.tsx to match feat/clippy-ai-el naming
- Drop re-exports of useAISearchContext/useChatContext from entry point
- Update callers to import directly: provider.tsx → ai-search,
  ai-search-trigger.tsx → chat-context, contact-form.tsx → chat-context
