+++ src/app/provider.tsx
@@ -5,7 +5,7 @@ import { TooltipProvider } from '@radix-ui/react-tooltip'
 import { NuqsAdapter } from 'nuqs/adapters/next/app'
 import type { ReactNode } from 'react'
 import { AISearchTrigger } from '@/components/ai/ai-search-trigger'
-import { AISearch } from '@/components/ai/chat'
+import { AISearch } from '@/components/ai/ai-search'
 import Analytics from '@/components/analytics'
 import { SmoothCursor } from '@/components/smooth-cursor'
 import { TailwindIndicator } from '@/components/tailwind-indicator'
